Save a ctd object to a netcdf file

Save a ctd object to a netcdf file

ctd2ncdf(x, varTable = NULL, ncfile = NULL, debug = 0)

Arguments

x

an oce object of class ctd, as created by e.g. oce::as.ctd() or oce::read.ctd().

varTable

character value indicating the variable-naming scheme to be used. This is provided to read.varTable(), the documentation of which explains the possibilities for pre-defined schemes named "argo" and "whp" and user-supplied files of similar format.

ncfile

character value naming the output file. Use NULL for a file name to be created automatically (e.g. ctd.nc for a CTD object).

debug

integer, 0 (the default) for quiet action apart from messages and warnings, or any larger value to see more output that describes the processing steps.

Author

Dan Kelley

Examples

library(ocencdf)

# example 1: a ctd file with no per-variable QC flags
data(ctd) # from 'oce' package
oce2ncdf(ctd, ncfile="ctd.nc")
#> Defaulting varTable to "argo".
#> Converting temperature from IPTS-68 scale to ITS-90 scale.
CTD <- as.ctd(ncdf2oce("ctd.nc"))
summary(CTD)
#> CTD Summary
#> -----------
#> 
#> * Start time:          2003-10-15 15:40:48
#> * Station:             Stn 2
#> * Mean Location:       44.684N 63.644W
#> * Data Overview
#> 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                   Min.   Mean   Max. Dim. NAs OriginalName
#>     scan           130    220    310  181   0         scan
#>     timeS          129    219    309  181   0        timeS
#>     pressure      1.48 22.885 44.141  181   0           pr
#>     depth        1.468 22.698 43.778  181   0         depS
#>     temperature 2.9183 7.5045 14.233  181   0         t068
#>     salinity    29.916 31.219 31.498  181   0        sal00
#>     flag             0      0      0  181   0         flag
#> 
#> * Processing Log
#> 
#>     - 2023-06-07 19:05:23 UTC: `create 'ctd' object`
#>     - 2023-06-07 19:05:23 UTC: `as.ctd(salinity = ncdf2oce("ctd.nc"))`

# example 2: a ctd file with per-variable QC flags
data(section) # from 'oce' package
stn <- section[["station", 100]] # 100-th station in section, not station '100'
oce2ncdf(stn, ncfile="stn.nc")
#> Defaulting varTable to "argo".
#> Converting temperature from IPTS-68 scale to ITS-90 scale.
STN <- as.ctd(ncdf2oce("stn.nc"))
summary(STN)
#> CTD Summary
#> -----------
#> 
#> * Start time:          1993-10-22 03:09:52
#> * Station:             109
#> * Mean Location:       36.323N 69.376W
#> * Data Overview
#> 
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#>                           Min.   Mean   Max. Dim. NAs OriginalName
#>     pressure              11.4 2182.5 4398.3   24   0       CTDPRS
#>     temperature         2.2228 7.3822 23.548   24   0       CTDTMP
#>     salinity            34.883 35.315 36.692   24   0       CTDSAL
#>     salinityBottle      34.889 35.315 36.702   24   0       SALNTY
#>     DOXY [μmol/kg]       144.9 246.91  282.1   24   0       OXYGEN
#>     silicate [μmol/kg]    1.71 16.853  36.52   24   0       SILCAT
#>     nitrite [μmol/kg]        0  0.022   0.08   24  19       NITRIT
#>     NO2+NO3 [μmol/kg]        0 14.619  23.73   24   0      NO2+NO3
#>     phosphate [μmol/kg]      0 0.8925   1.39   24   0       PHSPHT
#> 
#> * Data-quality Flags
#> 
#>     salinity:       "2" 22, "3" 2
#>     salinityBottle: "2" 24
#>     DOXY:           "2" 24
#>     silicate:       "2" 24
#>     nitrite:        "2" 5, "5" 19
#>     NO2+NO3:        "2" 24
#>     phosphate:      "2" 24
#> 
#> * Processing Log
#> 
#>     - 2023-06-07 19:05:23 UTC: `create 'ctd' object`
#>     - 2023-06-07 19:05:23 UTC: `as.ctd(salinity = ncdf2oce("stn.nc"))`

# clean up temporary files (to prevent CRAN test failure)
unlink("ctd.nc")
unlink("stn.nc")